refactor(console): keep sign-in method verification code check state
